requirements.txt
django==5.0.1
opencv-python==4.9.0.80
numpy==1.26.3
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==2.1.2+cu118
torchvision==0.16.2+cu118
torchaudio==2.1.2+cu118
pillow==10.2.0
scikit-learn==1.4.0
pandas==2.1.4
python-dotenv==1.0.0
django-tailwind==3.8.0
gunicorn==21.2.0
whitenoise==6.6.0
django-crispy-forms==2.1
crispy-tailwind==0.5.0
django-browser-reload==1.12.1
django-cors-headers==4.3.1
django-ratelimit==4.1.0
django-redis==5.4.0
redis==5.0.1
sentry-sdk==1.39.1
django-debug-toolbar==4.2.0
django-csp==3.7
django-axes==6.3.0
django-storages==1.14.2
django-cleanup==8.0.0
psutil==5.9.5
argon2-cffi==23.1.0
bcrypt==4.1.2
hiredis==2.3.2

# Production server
gunicorn==21.2.0
uvicorn==0.27.0
whitenoise==6.6.0
python-dotenv==1.0.0

# Monitoring and logging
sentry-sdk==1.39.1

# Windows WSGI server
waitress==2.1.2

# Enhanced training metrics and visualization
plotly>=5.18.0
seaborn>=0.13.0
torchviz>=0.0.2
scikit-learn>=1.3.2

# Training dependencies
tqdm>=4.66.1
albumentations>=1.4.0
graphviz>=0.20.3
matplotlib>=3.7.1
seaborn>=0.12.2
plotly>=5.13.0 